Added an option to filter projects by dependency.
Currently can show subgraph of all depending projects. Will add an
option for oposite (dependencies) later.
Improved how the graph is outputted and added support for project list
and GraphML output.
Work in progress on Dependency command which will be able to load
GraphML. It is close to be ready.